superbowl commercial sql project
This project aims to analyze Super Bowl commercials for 10 popular brands from 2000 to 2021. The study includes links to each commercial, along with additional information about the length, estimated cost, YouTube statistics, TV viewers, and defining characteristics. The analysis will be conducted using a combination of quantitative and qualitative methods to identify trends, patterns, and characteristics of the commercials. The goal of this project is to provide insights into the effectiveness of Super Bowl commercials and how they have evolved over time.